requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
pillow==11.2.1
python-docx==1.1.2
reportlab==4.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
